Functionality restored - needs cleanup +  refactor
