+++ src/app/feature/consultant/components/shared/consultant-skills-edit/consultant-skills-edit-reorder-display/consultant-skills-edit-reorder-display.component.scss
@@ -1,0 +1,48 @@
+/* Animate items as they're being sorted. */
+.cdk-drop-list-dragging .cdk-drag {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+/* Animate an item that has been dropped. */
+.cdk-drag-animating {
+    transition: transform 300ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.cdk-drag-preview {
+    box-sizing: border-box;
+    border-radius: 4px;
+    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
+                0 8px 10px 1px rgba(0, 0, 0, 0.14),
+                0 3px 14px 2px rgba(0, 0, 0, 0.12);
+}
+
+.display-skills {
+    margin-bottom: 1.5rem;
+    border-radius: 4px;
+    border: 1px solid rgba(0, 0, 0, 0.12);
+    overflow: hidden;
+}
+
+.display-skill {
+    align-items: center;
+    padding: .5rem;
+    border-bottom: 1px solid rgba(0, 0, 0, 0.12);
+    cursor: move;
+    display: block;
+    background: white;
+    box-sizing: border-box;
+}
+
+.display-skill-remove {
+    margin-right: .75rem;
+}
+
+.display-skill:last-child {
+    border: none;
+}
+
+.subtitle {
+    color: rgba(0,0,0,.54);
+    margin-bottom: 1.5rem;
+    margin-top: -1rem;
+}
