--- src/app/feature/consultant/components/shared/consultant-skills-edit/consultant-skills-edit.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ConsultantSkillsEditService {
-
-  constructor() { }
-}
